requirements.txt
appdirs==1.4.0
cycler==0.10.0
dask==0.13.0
decorator==4.0.11
Django==1.10.5
django-crispy-forms==1.6.1
django-environ==0.4.1
django-flat-responsive==1.2.0
django-model-utils==2.6.1
google-api-python-client==1.6.2
httplib2==0.10.3
matplotlib==2.0.0
networkx==1.11
numpy==1.12.0+mkl
oauth2client==4.0.0
olefile==0.44
opencv-python==3.2.0
packaging==16.8
Pillow==4.0.0
pyasn1==0.2.2
pyasn1-modules==0.0.8
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
requests==2.13.0
rsa==3.4.2
scikit-image==0.12.3
six==1.10.0
standardjson==0.3.1
toolz==0.8.2
uritemplate==3.0.0